add question popup when delete member, fix update detail

+++ MyStoreWinApp/frmMemberManagement.cs
@@ -51,9 +51,12 @@ namespace MyStoreWinApp
         {
             try
             {
-                var member = GetMemberObject();
-                memberRepo.DeleteMember(member.MemberID.ToString());
-                LoadMembersToGridView(memberRepo.GetAllMembers());
+                if(MessageBox.Show("Are you sure to delete?", "Confirm", MessageBoxButtons.YesNo) == DialogResult.Yes)
+                {
+                    var member = GetMemberObject();
+                    memberRepo.DeleteMember(member.MemberID.ToString());
+                    LoadMembersToGridView(memberRepo.GetAllMembers());
+                }
             }
             catch (Exception ex)
             {
